requirements.txt
spotipy>=2.23.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
python-dotenv>=1.0.0
jupyter>=1.0.0
streamlit>=1.28.0
plotly>=5.15.0
tqdm>=4.65.0
black>=23.0.0         
pytest>=7.0.0        